Feat: Remove root archive file system provider
Now that ArchiveReader supports SeekableByteChannel.

Fixes: #1140
